api/requirements.txt
flask
flask-cors >= 3.0.3
flask-restful >= 0.3.6
python-dotenv >= 0.6.4
python_dateutil == 2.6.0
sqlalchemy
